documentation

+++ src/main/java/org/kristiania/smartinventorymanagementsystem/controllers/InventoryController.java
@@ -4,8 +4,9 @@ import org.kristiania.smartinventorymanagementsystem.model.Product;
 import org.kristiania.smartinventorymanagementsystem.service.InventoryService;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
-
+/**
+ * Handles endpoints for inventory management requests
+ */
 @RestController
 @RequestMapping("/api/inventory")
 public class InventoryController {
@@ -16,13 +17,22 @@ public class InventoryController {
         this.inventoryService = inventoryService;
     }
 
+    /**
+     * Handles the endpoint for adding products to database
+     * @param product to add to database
+     */
     @PostMapping("/products")
     public void addProduct(@RequestBody Product product) {
         inventoryService.addProduct(product);
     }
 
+    /**
+     * Handles endpoint for selling products
+     * @param id of sold product
+     * @param quantity sold of product
+     */
     @PostMapping("/products/{id}/sell")
-    public void sellProduct(@PathVariable Long id,
+    public void sellProduct(@PathVariable int id,
                             @RequestParam int quantity) {
         inventoryService.sellProduct(id, quantity);
     }
