+++ src/main/java/org/kristiania/smartinventorymanagementsystem/model/AbstractUser.java
@@ -7,17 +7,18 @@ import lombok.Setter;
 @Getter
 public abstract class AbstractUser {
     // Getters and setters
-    private Long id;
+    private int id;
     private String username;
     private String email;
 
     public AbstractUser() {}
 
-    public AbstractUser(Long id, String username, String email) {
+    public AbstractUser(int id, String username, String email) {
         this.id = id;
         this.username = username;
         this.email = email;
     }
 
     // Possibly add common methods like authenticate(), etc.
+
 }
