+++ src/main/java/org/kristiania/smartinventorymanagementsystem/model/Customer.java
@@ -3,7 +3,7 @@ package org.kristiania.smartinventorymanagementsystem.model;
 public class Customer extends AbstractUser {
     public Customer() {}
 
-    public Customer(Long id, String username, String email) {
+    public Customer(int id, String username, String email) {
         super(id, username, email);
     }
